pyproject.toml
[project]
name = "s3-json-to-parquet-rs"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "marioweid", email = "mario.weidner@gmx.de" }
]
dependencies = [
    "polars>=1.8.2",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/s3_json_to_parquet_rs"]
